Replace Web3 package with a custom minimal web3 client implementation Minor improvements
